requirements.txt
echo "Django==4.2.3
django-cors-headers==3.13.0
requests==2.31.0
urllib3==1.26.10
whitenoise==6.2.0
psycopg2-binary==2.9.3
dj-database-url==1.0.0
django-pwa==2.0.1
django-storages==1.14.0
boto3==1.34.0
gunicorn==20.1.0
intasend==0.2." > requirements.txt
